requirements.txt
# YouTube Automation Requirements

# Core dependencies
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
google-auth>=2.0.0

# Media processing
moviepy>=1.0.3
Pillow>=9.0.0
gtts>=2.3.0
pydub>=0.25.1

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
beautifulsoup4>=4.11.0
schedule>=1.1.0

